Split the logic between api / events / DOM to make it more readable and maintanable
