+++ lib/cjs/services/dom.d.ts
@@ -1,0 +1,2 @@
+export declare let updateHTML: (id: string, update: string) => void;
+export declare let showJoke: () => Promise<void>;
